Refactor fetch and rendor data function
