+++ src/index.js
@@ -1,71 +1,41 @@
-// Import the functions you need
-import { initializeApp } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-app.js";
-import { getFirestore, collection, addDoc, getDocs } from "https://www.gstatic.com/firebasejs/10.3.0/firebase-firestore.js";
+import { fetchData } from "./firestore.js";
 
-// Firebase config
-const firebaseConfig = {
-    apiKey: "AIzaSyC2560pETTPekmI-ABPf_xZ3M3UB6ZqMK4",
-    authDomain: "message-display-system.firebaseapp.com",
-    projectId: "message-display-system",
-    storageBucket: "message-display-system.appspot.com",
-    messagingSenderId: "260068152843",
-    appId: "1:260068152843:web:2336283608fddc0e41d90e",
-    measurementId: "G-HCK0JW6RYC"
-};
+async function render(event) {
+  const tableBody = document.getElementById("eventTableBody");
+  const row = document.createElement("tr");
 
-// Initialize Firebase and Firestore
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
+  const organiserCell = document.createElement("td");
+  organiserCell.textContent = event.organiserName;
+  row.appendChild(organiserCell);
 
-// General function to add data to Firestore
-async function addDataToFirestore(collectionName, data) {
-  try {
-    await addDoc(collection(db, collectionName), data);
-    console.log("Data successfully added!");
-  } catch (e) {
-    console.error("Error adding data: ", e);
-  }
+  const jsDate = event.time.toDate();
+  const dateCell = document.createElement("td");
+  const formattedDate = jsDate.toLocaleDateString();
+  dateCell.textContent = formattedDate;
+  row.appendChild(dateCell);
+
+  const timeCell = document.createElement("td");
+  const formattedTime = jsDate.toLocaleTimeString();
+  timeCell.textContent = formattedTime;
+  row.appendChild(timeCell);
+
+  const wifiCell = document.createElement("td");
+  wifiCell.textContent = event.wifi;
+  row.appendChild(wifiCell);
+
+  const categoryCell = document.createElement("td");
+  categoryCell.textContent = event.category;
+  row.appendChild(categoryCell);
+
+  tableBody?.appendChild(row);
 }
 
-// Usage example:
-// const inputData = { key1: "value1", key2: "value2" };
-// addDataToFirestore("yourCollectionName", inputData);
-
-// fetch data from Firestore
-async function fetchData() {
-  const querySnapshot = await getDocs(collection(db, "events"));
-  const tableBody = document.getElementById("eventTableBody");
-  
-  querySnapshot.forEach((doc) => {
-    const event = doc.data();
-    const row = document.createElement("tr");
-
-    const organiserCell = document.createElement("td");
-    organiserCell.textContent = event.organiserName;
-    row.appendChild(organiserCell);
-
-    const jsDate = event.time.toDate();
-    const dateCell = document.createElement("td");
-    const formattedDate = jsDate.toLocaleDateString();
-    dateCell.textContent = formattedDate;
-    row.appendChild(dateCell);
-
-    const timeCell = document.createElement("td");
-    const formattedTime = jsDate.toLocaleTimeString();
-    timeCell.textContent = formattedTime;
-    row.appendChild(timeCell);
-
-    const wifiCell = document.createElement("td");
-    wifiCell.textContent = event.wifi;
-    row.appendChild(wifiCell);
-
-    const categoryCell = document.createElement("td"); 
-    categoryCell.textContent = event.category; 
-    row.appendChild(categoryCell); 
-
-    tableBody.appendChild(row);
+async function fetchDataAndRender() {
+  const events = await fetchData();
+  events.forEach((event) => {
+    render(event);
   });
 }
 
-// Fetch the data
-fetchData();
+// Fetch the data and render
+fetchDataAndRender();
